+++ engine/h2shared/gl_draw.c
@@ -574,7 +574,7 @@ void Draw_SmallCharacter (int x, int y, int num)
 	if (num == 0)
 		return;
 
-	if (y <= -8 || y >= vid.height)
+	if (y <= -8 || y >= (int)vid.height)
 		return; 	// totally off screen
 
 	row = num >> 4;
@@ -839,7 +839,7 @@ void Draw_PicCropped (int x, int y, qpic_t *pic)
 	glpic_t 	*gl;
 	float		th, tl;
 
-	if ((x < 0) || (x+pic->width > vid.width))
+	if ((x < 0) || (x+pic->width > (int)vid.width))
 	{
 		Sys_Error("%s: bad coordinates", __thisfunc__);
 	}
@@ -851,7 +851,7 @@ void Draw_PicCropped (int x, int y, qpic_t *pic)
 
 	// rjr	tl/th need to be computed based upon pic->tl and pic->th
 	//	cuz the piece may come from the scrap
-	if (y+pic->height > vid.height)
+	if (y+pic->height > (int)vid.height)
 	{
 		height = vid.height-y;
 		tl = 0;
@@ -892,7 +892,7 @@ void Draw_SubPicCropped (int x, int y, int h, qpic_t *pic)
 	glpic_t 	*gl;
 	float		th,tl;
 
-	if ((x < 0) || (x+pic->width > vid.width))
+	if ((x < 0) || (x+pic->width > (int)vid.width))
 	{
 		Sys_Error("%s: bad coordinates", __thisfunc__);
 	}
@@ -904,7 +904,7 @@ void Draw_SubPicCropped (int x, int y, int h, qpic_t *pic)
 
 	// rjr	tl/th need to be computed based upon pic->tl and pic->th
 	//	cuz the piece may come from the scrap
-	if (y+pic->height > vid.height)
+	if (y+pic->height > (int)vid.height)
 	{
 		height = vid.height-y;
 		tl = 0;
@@ -1177,9 +1177,9 @@ void Draw_FadeScreen (void)
 			bx = 0;
 		if (by < 0)
 			by = 0;
-		if (ex > vid.width)
+		if (ex > (int)vid.width)
 			ex = vid.width;
-		if (ey > vid.height)
+		if (ey > (int)vid.height)
 			ey = vid.height;
 
 		glBegin_fp (GL_QUADS);
